requirements.txt
Django==5.0.6
djangorestframework==3.15.1
django-cors-headers==4.3.1
python-decouple==3.8
psycopg[binary]==3.2.3
pgvector==0.3.4
openai==1.51.2
# langchain==0.3.1
# langchain-openai==0.2.1
# langchain-community==0.3.1
celery==5.4.0
redis==5.1.1
gunicorn==23.0.0
whitenoise==6.8.1
cloudinary==1.41.0
numpy>=1.26.0
scikit-learn>=1.5.0
dj-database-url==2.2.0
django-filter==24.3